pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uponu_dc_fab_inventory"
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "pynetbox",
    "deepmerge>=1.1.0",
    "PyYAML>=6.0.0",
    "click>=8.1"
]

[project.optional-dependencies]

dev = [
  "pylint-pydantic>=0.2.4",
  "pylint>=2.17.5",
  "pytest-asyncio>=0.21.1",
  "pytest-cov>=4.1.0",
  "pytest-dependency",
  "pytest-html>=3.2.0",
  "pytest-metadata>=3.0.0",
  "pytest>=7.4.0",
]

[project.scripts]

upnetfabinv = "uponu_dc_fab_inventory.cli:cli"